requirements.txt
streamlit
moviepy
Pillow
trimesh
svgwrite
openai
